+++ user_agent_blacklist.go
@@ -2,8 +2,11 @@ package pirsch
 
 // userAgentBlacklist contains all substrings (in lowercase) used to filter the User-Agent header.
 // Please add the reference in case you copy an existing list.
-// - https://github.com/gorangajic/isbot/blob/master/list.json
 var userAgentBlacklist = []string{
+	// custom
+	"saashub",
+
+	// https://github.com/gorangajic/isbot/blob/master/list.json
 	"://", // urls
 	"bot",
 	"crawler",